Made minor changes to all files.
Building class now successfully takes in number of floors and creates
all floors and stores it in an ArrayList.
Added a few private variables to elevator class to help with determining
weight. Also put in accessor methods.
Almost completed Floor class. Just need to check if methods work
correctly and need elevator class to work to test.
Made Generator class output time first because that is needed to make a
person before we can use the starting and ending points.
Added weight variable to person class.